# src/main/java/com/prx/backoffice/v1/people/api/controller/PersonController.java
/*
 *  @(#)PersonController.java
 *
 *  Copyright (c) Luis Antonio Mata Mata. All rights reserved.
 *
 *   All rights to this product are owned by Luis Antonio Mata Mata and may only
 *  be used under the terms of its associated license document. You may NOT
 *  copy, modify, sublicense, or distribute this source file or portions of
 *  it unless previously authorized in writing by Luis Antonio Mata Mata.
 *  In any event, this notice and the above copyright must always be included
 *  verbatim with this file.
 */
package com.prx.backoffice.v1.people.api.controller;

import com.prx.backoffice.util.MessageUtil;
import com.prx.backoffice.v1.people.api.to.PersonRequest;
import com.prx.backoffice.v1.people.service.PersonService;
import com.prx.commons.general.pojo.Person;
import com.prx.commons.general.to.Response;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.UUID;

/**
 * PersonController. Clase controladora para la exposición de los endpoint pertenecientes a la gestión de usuario
 *
 * @author Luis Antonio Mata
 * @version 1.0.1.20200904-01, 04-11-2020
 */
@RestController
@CrossOrigin(origins = "*")
@RequestMapping("/api/v1/people")
public class PersonController {
    /** personService */
    private final PersonService personService;

    public PersonController(PersonService personService) {
        this.personService = personService;
    }

    /**
     *
     * @param personRequest {@link PersonRequest}
     * @return Objeto de tipo {@link Response}
     */
    @Operation(summary = "Create a person", description = "Create and return the person.")
    @ApiResponses(value = {
        @ApiResponse(responseCode = MessageUtil.OK, description = "Person created")
    })
    @PostMapping(produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE, path = "/")
    public ResponseEntity<Person> create(@Parameter(description = "Request to create a person", required = true)
    @RequestBody final PersonRequest personRequest) {
        return personService.create(personRequest.getPerson());
    }

    @Operation(summary = "Fetch person by ID", description = "Return a person.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = MessageUtil.OK, description = "OK"),
            @ApiResponse(responseCode = MessageUtil.NOT_FOUND, description = "Person not found"),
            @ApiResponse(responseCode = MessageUtil.UNPROCESSABLE_ENTITY, description = "Person not found")
    })
    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE, path = "/{personId}")
    public ResponseEntity<Person> find(@PathVariable final UUID personId) {
        return personService.find(personId);
    }

    @Operation(summary = "Update a person", description = "Update and return the person.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = MessageUtil.OK, description = "Person updated"),
            @ApiResponse(responseCode = MessageUtil.BAD_REQUEST, description = "PersonId invalid"),
            @ApiResponse(responseCode = MessageUtil.BAD_REQUEST, description = "Person request invalid"),
            @ApiResponse(responseCode = MessageUtil.BAD_REQUEST, description = "Person not founded")
    })
    @PutMapping(produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE, path = "/{personId}")
    public ResponseEntity<Person> update(@PathVariable final UUID personId, @RequestBody final PersonRequest personRequest) {
        return personService.update(personId, personRequest.getPerson());
    }

    @Operation(summary = "Fetch all people", description = "Return all people.")
    @ApiResponses(value = {
            @ApiResponse(responseCode = MessageUtil.OK, description = "OK"),
            @ApiResponse(responseCode = MessageUtil.NOT_FOUND, description = "Person not found")
    })
    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE, path = "/")
    public ResponseEntity<List<Person>> list() {
        return personService.list((UUID) null);
    }
}