# src/main/resources/META-INF/backbone_rest-openapi.yaml
openapi: "3.1.0"
info:
  title: "Backbone REST API"
  description: "backbone_rest API"
  version: "1.0.0"
servers:
- url: "https://backbone_rest"
...
paths:
...
  /api/v1/contacts/person/{personId}:
    get:
      summary: ""
      description: "List contacts by person Id."
      operationId: ""
      parameters:
      - name: "personId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Contact"
...
  /api/v1/people/{personId}:
    put:
      summary: "Update a person"
      description: "Update and return the person."
      operationId: ""
      parameters:
      - name: "personId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
    get:
      summary: "Fetch person by ID"
      description: "Return a person."
      operationId: ""
      parameters:
      - name: "personId"
        in: "path"
        required: true
        schema:
          type: "string"
          format: "uuid"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
  /api/v1/people/:
    post:
      summary: "Create a person"
      description: "Create and return the person."
      operationId: ""
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
    get:
      summary: "Fetch all people"
      description: "Return all people."
      operationId: ""
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Person"
  /api/v1/report/template:
    get:
      summary: "GET api/v1/report/template"
      operationId: "createWordDocument"
      parameters:
      - name: "values"
        in: "query"
        required: true
        schema:
          type: "string"
      - name: "documentTemplate"
        in: "query"
        required: true
        schema:
          type: "string"
          format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            application/octet-stream:
              schema:
                $ref: "#/components/schemas/Resource"
  /api/v1/report/placeholdervalues:
    get:
      summary: "GET api/v1/report/placeholdervalues"
      operationId: "placeholderValues"
      parameters:
      - name: "templateDocumentModel"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/TemplateDocumentModel"
      - name: "documentTemplate"
        in: "query"
        required: true
        schema:
          type: "string"
          format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            application/octet-stream:
              schema:
                type: "array"
                items:
                  type: "string"
...
components:
  schemas:
    ContactType:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        name:
          type: "string"
        description:
          type: "string"
        active:
          type: "boolean"
    Person:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        firstName:
          type: "string"
        middleName:
          type: "string"
        lastName:
          type: "string"
        gender:
          type: "string"
        birthdate:
          type: "string"
          format: "date"
    Contact:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        content:
          type: "string"
        contactType:
          $ref: "#/components/schemas/ContactType"
        person:
          $ref: "#/components/schemas/Person"
        active:
          type: "boolean"
...
    PersonRequest:
      type: "object"
      properties:
        person:
          $ref: "#/components/schemas/Person"
        dateTime:
          type: "string"
          format: "date-time"
        appName:
          type: "string"
        appToken:
          type: "string"
    Resource:
      type: "object"
      properties: {}
    TemplateDocumentModel:
      type: "object"
      properties:
        versionId:
          type: "integer"
          format: "int64"
        templateName:
          type: "string"
        description:
          type: "string"
        createdBy:
          type: "string"
        createdTimestamp:
          type: "string"
          format: "date"
        lastModifiedBy:
          type: "string"
        lastModifiedTimestamp:
          type: "string"
          format: "date"
...